bugfix: add initial bind port check
Assuming there is a "nc" process run before zjuconnect, which listening same tcp/udp port. By rights zjuconnect need report this port is already in use and exit, however due to "nc" use "SO_REUSEPORT" and "SO_REUSEADDR", zjuconnect still can bind to this port successfully, however the behavior for all sockets bound to that port is indeterminate (https://learn.microsoft.com/en-us/windows/win32/winsock/using-so-reuseaddr-and-so-exclusiveaddruse#using-so_reuseaddr). So we need to check for any potential port conflicts during initialization.

Be attention that we use a strict policy to check, for example assume nc listen to local TCP addr "192.168.189.123:1234", and we want to listen to 127.0.0.1:1234, we also prohibit this operation although this operation maybe legal.